% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly-plot.R
\name{monthly_plot}
\alias{monthly_plot}
\title{Plot Average Monthly Measles Cases by Region}
\usage{
monthly_plot(
  selected_region,
  selected_months = NULL,
  data = load_data()$cases_month
)
}
\arguments{
\item{selected_region}{A character string specifying the WHO region code.
It can be one of the following listed: "AFR", "AMR", "EMR", "EUR", "SEAR", "WPR".}

\item{selected_months}{A numeric vector that can take month numbers between 1-12.
If no months are chosen specifically, and its NULL then all months are included.}

\item{data}{A data frame containing monthly measles case data, defaulted to cases_month tibble on monthly measles case data.}
}
\value{
A ggplot object showing average monthly measles cases.
}
\description{
Creates a line plot of average monthly measles cases for a chosen WHO
region, with an optional filter for specific months.
}
\examples{
cases_month <- load_data()$cases_month
monthly_plot("AMR")
monthly_plot("EUR", selected_months = c(1, 2, 3))
}
