% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly-table.R
\name{monthly_table}
\alias{monthly_table}
\title{Summarize Average Monthly Measles Cases by Region}
\usage{
monthly_table(data, selected_region, selected_months = NULL)
}
\arguments{
\item{data}{A data frame containing monthly measles case data.}

\item{selected_region}{A character string specifying the WHO region code.
One of "AFR", "AMR", "EMR", "EUR", "SEAR", "WPR".}

\item{selected_months}{A numeric vector of month numbers from 1-12.
If no months are chosen then all months are included.}
}
\value{
A gt table of average monthly measles cases for the selected region.
}
\description{
Returns a gt table of average monthly measles cases for chosen WHO
region, with an optional filter for specific months.
}
\examples{
monthly_table(cases_month, "AMR")
monthly_table(cases_month, "EUR", selected_months = c(1, 2, 3))
}
